feat: tariff-watch — weekly automated price-seed updates
tools/tariff_watch: deterministic parsers for the official sources
(Galp PDF, Endesa Digital annex incl. word-coordinate reconstruction of
its zipped gas rows, Goldenergy reference page, regulated gas table)
that patch YAMLs in place preserving comments and never rewrite equal
values; a price-token fingerprint watcher for comparator sources whose
URL list derives from the seeds' own source_url fields; weekly workflow
that opens a reviewable PR for parser changes and deduped issues for
watch hits and parse failures. Live run against all sources reproduces
the audited seeds exactly. Also: CI no longer installs the removed
playwright, and the four long-standing lint findings are fixed — ruff
is fully clean for the first time.

+++ portinhola/api/calculator.py
@@ -253,15 +253,13 @@ def potencia(supply_point_id: int, db: Session = Depends(get_db)) -> dict:
             from portinhola.core.replay import _nearest_power_price
 
             prices = current_tariff.electricity.power_eur_day
-            saving = int(
-                round(
-                    (
-                        _nearest_power_price(prices, contracted)
-                        - _nearest_power_price(prices, recommended)
-                    )
-                    * 365.25
-                    * 100
+            saving = round(
+                (
+                    _nearest_power_price(prices, contracted)
+                    - _nearest_power_price(prices, recommended)
                 )
+                * 365.25
+                * 100
             )
 
     return {
